Clear the findings that stand in the way of linting in CI
Preparation for switching on CI checks: every linter that is worth running
here reports findings on the current tree, so a workflow added today would be
red from the first run. This clears them, so the workflow can start green and
stay meaningful.

Shell (shellcheck -x -S warning, 11 -> 0)
- phobos-common.sh: quote the array subscripts passed to unset, otherwise a
  path containing glob characters would be expanded before removal.
- phobos-common.sh is a library, so every variable it defines is read by the
  scripts that source it and never here. SC2034 is suppressed once for the
  whole file, with the reason stated, instead of twelve separate directives.
- phobos-filesystem.sh: read the tail flags line by line into an array. The
  word splitting is intended, but now it is explicit and survives a
  multi-line file.
- phobos_wrapper.sh: drop rlimit_arg. It was assembled but never used;
  "ulimit -v" a few lines further down does the actual limiting.

C (cppcheck, 1 -> 0)
- netcache_probe.c: the conditional promotes unsigned short to int, so "%u"
  received a signed value. Cast explicitly.

Python (ruff 62 -> 0, bandit 1 high -> 0)
- Applied ruff's automatic fixes (import splitting and ordering, PEP 585
  annotations).
- orchestrate.py: run() no longer accepts a command string and never uses a
  shell. Both call sites pass argument lists, so the string branch was dead
  code that only widened the injection surface.
- The two broad "except Exception" clauses are deliberate and stay: one keeps
  a single failing language from aborting the other prune runs, the other is
  the fallback when the external realpath fails. Both are now marked with the
  reason.
- The three scripts carry a shebang and are now executable.

YAML (yamllint 13 -> 0) and Docker (hadolint, warnings -> 0)
- Two files had CRLF line endings, which breaks tooling on Linux.
- .yamllint relaxes the rules that buy nothing here (document start, 80
  columns) but pins line endings to LF.
- .hadolint.yaml ignores DL3008 with the reason: Ubuntu keeps only the current
  version of a package in its archive, so pinning apt versions breaks the
  build as soon as a security update ships. The base image tags are pinned
  instead and Renovate keeps them current.
- .bandit skips B108: the /var/tmp paths are the agreed mount points inside
  the containers, not temporary files, and they are overridable defaults.

No behaviour change intended. tests/timeout_units.sh (58 passed) and
tests/network_cache_ports.sh (20 passed) both still pass.

+++ tests/netcache_probe.c
@@ -177,7 +177,8 @@ static int phase_one(const char *self, const char *scenario) {
     argv[n++] = args[i];
   }
   for (int i = 0; i < PORTS_V6; i++) {
-    snprintf(args[PORTS_V4 + i], sizeof args[PORTS_V4 + i], "%u", have_v6 ? v6[i].port : 0);
+    snprintf(args[PORTS_V4 + i], sizeof args[PORTS_V4 + i], "%u",
+             (unsigned int) (have_v6 ? v6[i].port : 0));
     argv[n++] = args[PORTS_V4 + i];
   }
   argv[n] = NULL;
